web/requirements.txt
flask
requests
python-dotenv
boto3
flask-cors
gunicorn